requirements.txt
absl-py==1.4.0
array-record==0.2.0
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cycler==0.11.0
Cython==0.29.34
dataclasses==0.6
decorator==5.1.1
dm-tree==0.1.8
etils==1.2.0
fire==0.5.0
flatbuffers==23.3.3
gast==0.4.0
gin-config==0.5.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
h5py==3.8.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
jax==0.4.8
joblib==1.2.0
kaggle==1.5.13
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
librosa==0.8.1
llvmlite==0.36.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.4.3
ml-dtypes==0.1.0
neural-structured-learning==1.4.0
numba==0.53.0
numpy==1.24.2
oauthlib==3.2.2
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
platformdirs==3.2.0
pooch==1.7.0
promise==2.3
proto-plus==1.22.2
protobuf==4.22.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycocotools==2.0.6
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scann==1.2.6
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.98
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-addons==0.20.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
tensorflow-model-optimization==0.7.4
tensorflowjs==3.18.0
termcolor==2.2.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.3.0
tf-slim==1.1.0
tflite-model-maker==0.4.2
tflite-support==0.4.3
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.65.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.25.11
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
